requirements.txt
beautifulsoup4==4.13.3
langchain_community==0.3.17
langchain_core==0.3.34
langchain_huggingface==0.1.2
langchain_openai==0.3.4
langchain_text_splitters==0.3.6
python-dotenv==1.0.1
Requests==2.32.3
torch==2.6.0
